refactor: use simd as default feature
